Arrow function and when not to use?
Never use arrow functions as methods in objects, prototype methods, class methods - as this inside arrow function refers to its parent - parent would be window object

arrow functions can be used in callback chaining, promises chaining and array methods
This in arrow function?

this in regular function refers to object using which function is called. this refers to function execution context
this in arrow function refers to its parent or where the function is defined. It doesn't depend on the calling object. In arrow function this will refer to global object unless and untill used inside inner function enclosing outer function which has this defined - in this case, it refers to parent -> outer function